feat(embeddable-tcp-server): generalize tcp worker bridge
