+++ code/packages/rust/Cargo.toml
@@ -341,6 +341,6 @@ members = [
     "state-machine-markup-serializer",
     "state-machine-markup-deserializer",
     "state-machine-markup-json-serializer",
-    "python-mini-redis-tcp",
+    "embeddable-tcp-server",
     "generic-job-protocol",
 ]
