+++ code/packages/rust/embeddable-tcp-server/BUILD
@@ -1,0 +1,1 @@
+cargo test -p embeddable-tcp-server -- --nocapture
