+++ code/packages/rust/embeddable-tcp-server/Cargo.toml
@@ -1,12 +1,14 @@
 [package]
-name = "python-mini-redis-tcp"
+name = "embeddable-tcp-server"
 version = "0.1.0"
 edition = "2021"
-description = "Rust TCP runtime bridge that delegates Mini Redis commands to Python workers"
+description = "Language-neutral embeddable TCP server for generic job-protocol workers"
 
 [dependencies]
 generic-job-protocol = { path = "../generic-job-protocol" }
-resp-protocol = { path = "../resp-protocol" }
 serde = { version = "1", features = ["derive"] }
 tcp-runtime = { path = "../tcp-runtime" }
 transport-platform = { path = "../transport-platform" }
+
+[dev-dependencies]
+resp-protocol = { path = "../resp-protocol" }
