--- code/packages/rust/python-mini-redis-tcp/BUILD
@@ -1,1 +1,0 @@
-cargo test -p python-mini-redis-tcp -- --nocapture
