--- code/packages/rust/python-mini-redis-tcp/required_capabilities.json
@@ -1,26 +1,0 @@
-{
-  "$schema": "https://raw.githubusercontent.com/adhithyan15/coding-adventures/main/code/specs/schemas/required_capabilities.schema.json",
-  "version": 1,
-  "package": "rust/python-mini-redis-tcp",
-  "capabilities": [
-    {
-      "category": "net",
-      "action": "listen",
-      "target": "127.0.0.1:*",
-      "justification": "The prototype opens a local TCP listener through tcp-runtime so clients can speak RESP over real sockets."
-    },
-    {
-      "category": "proc",
-      "action": "exec",
-      "target": "python*",
-      "justification": "The prototype launches the Python Mini Redis worker as a child process to prove the cross-language runtime seam."
-    },
-    {
-      "category": "env",
-      "action": "read",
-      "target": "PYTHON",
-      "justification": "Tests may read PYTHON to select an explicit Python 3.12 interpreter for the worker process."
-    }
-  ],
-  "justification": "This crate intentionally owns the TCP listener and child worker process for the Python Mini Redis over Rust TCP runtime prototype."
-}
